# 8306706: Support out-of-line code generation for MachNodes