Hi,

This patch adds supports for MachNodes to emit an out-of-line piece of code in the stub section of the compiled method. This allows the separation of the uncommon path from the common one, which speeds up the common path a little bit and increases compiled code density. Please take a look and leave reviews.

Thanks a lot.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8306706](https://bugs.openjdk.org/browse/JDK-8306706): Support out-of-line code generation for MachNodes


### Reviewers
 * [Tobias Hartmann](https://openjdk.org/census#thartmann) (@TobiHartmann - **Reviewer**) ⚠️ Review applies to [a17bcb76](https://git.openjdk.org/jdk/pull/13602/files/a17bcb76bac7bb207728fbe73a80eae3087b5c1c)
 * [Vladimir Kozlov](https://openjdk.org/census#kvn) (@vnkozlov - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/13602/head:pull/13602` \
`$ git checkout pull/13602`

Update a local copy of the PR: \
`$ git checkout pull/13602` \
`$ git pull https://git.openjdk.org/jdk.git pull/13602/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 13602`

View PR using the GUI difftool: \
`$ git pr show -t 13602`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/13602.diff">https://git.openjdk.org/jdk/pull/13602.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/13602#issuecomment-1519127252)