@merykitty 2023-04-23
The generated node for the stub looks like this:

    class convF2I_reg_regStub : public C2CodeStub {
    private:
        const convF2I_reg_regNode* _node;
        PhaseRegAlloc* ra_;
        MachOper* opnd_array(uint index) const { return _node->opnd_array(index); }
    public:
        convF2I_reg_regStub(const convF2I_reg_regNode* node, PhaseRegAlloc* ra) : _node(node), ra_(ra) {}
        int max_size() const { return 23; }
        void emit(C2_MacroAssembler& masm);
    };

And the corresponding node's `emit` method has an additional section:

    void convF2I_reg_regNode::emit(CodeBuffer& cbuf, PhaseRegAlloc* ra_) const {
        cbuf.set_insts_mark();
        convF2I_reg_regStub* stub = new (Compile::current()->comp_arena()) convF2I_reg_regStub(this, ra_);
        if (!Compile::current()->output()->in_scratch_emit_size()) {
            Compile::current()->output()->add_stub(stub);
        }
        ...
    }