@merykitty 2023-05-17
@TobiHartmann Thanks for taking a look, I think this can be used for the vectorized version of these nodes, as well as the max, min nodes for floating point numbers. I also see compact header uses out-of-line code to slow path `LoadNKlass`.