@vnkozlov 2023-05-17
> @vnkozlov Yes we can explicitly define a stub without relying on code generation, it may be more preferable since it avoids adding complexity to adlc generation. The only downside is that there is some boilerplate for each usage but I think the boilerplate is not too terrible.

Can you look on that? There could be other cases in Macroassembler which can use this