@merykitty 2023-05-23
@vnkozlov Thanks for your reviews and testing
@TobiHartmann Thanks for your suggestion, I have added comments to describe the purpose of `C2GeneralStub`